+++ lib/ajax_pagination/rails.rb
@@ -14,5 +14,13 @@ module AjaxPagination
         AjaxPagination.warnings = Rails.env == 'development' # default setting
       end
     end
+
+    initializer 'ajax_pagination.redirect_filter' do
+      if AjaxPagination.redirect_after_filter == true
+        ActiveSupport.on_load(:action_controller) do
+          after_filter :ajax_pagination_redirect
+        end
+      end
+    end
   end
 end