+++ spec/rails_app/app/controllers/application_controller.rb
@@ -1,8 +1,7 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery
   before_filter :slowajaxload
-  ajax_pagination :pagination => 'menu'
-  #ajax_pagination :pagination => 'sdf', :partial => {:template => "pages/readme"}
+  ajax_pagination :pagination => 'menu', :render => { :layout => "flash" }
   def slowajaxload
     if params[:pagination] && Rails.env == "test"
       sleep(0.5)